requirements.txt
accelerate==1.11.0
bitsandbytes==0.48.2
cairosvg
captum
cython
datasets==4.3.0
einops
evaluate==0.4.6
ipykernel
kaleido
matplotlib
ninja
packaging
peft==0.17.1
pillow
plotly
protobuf==6.33.0
qwen-vl-utils
scikit-learn
scikit-image
torch==2.9.0
torchvision==0.24.0
transformers==4.57.1
wandb